examples/yolov3_embeddings/requirements.txt
numpy
torch >= 1.8.1
torchvision >= 0.8.2
terminaltables == 3.1.0
Pillow == 8.1.0
tqdm == 4.55.1
imgaug == 0.4.0
torchsummary == 1.5.1